requirements.txt
pandas
numpy
tefas-crawler
openpyxl
xlsxwriter
ruff